requirements.txt
tensorflow
keras
tensorflow-datasets
matplotlib
discord.py[voice]
lyricsgenius
youtube-dl
selenium
webdriver_manager
pytest
distest
